requirements.txt
streamlit
google-generativeai
speechrecognition
pydub